[core] Support re-overwrite by spark for chain table
